購入履歴削除機能

+++ TeamBProject/src/main/java/ec/com/services/LessonService.java
@@ -13,6 +13,7 @@ import ec.com.model.dao.TransactionHistoryDao;
 import ec.com.model.dao.TransactionItemDao;
 import ec.com.model.dto.LessonWithTransactionDto;
 import ec.com.model.entity.Lesson;
+import ec.com.model.entity.TransactionItem;
 import jakarta.transaction.Transactional;
 
 @Service
@@ -126,15 +127,40 @@ public class LessonService {
 		        dto.setLessonFee((Integer) row[7]);
 		        dto.setTransactionDate(((Timestamp) row[8]).toLocalDateTime().toLocalDate());
 		        dto.setTransactionId(((Number) row[9]).longValue());
+		        dto.setId(((Number) row[10]).longValue());
 		        list.add(dto);
 		    }
 		return list;
 	}
 	
-	// 購入履歴の削除
+	/**
+	 *  購入履歴の削除を行う。
+	 *  選択されたitemをDBから削除処理
+	 *  その後、transactionHistoryを削除する。
+	 *  transactionHistoryはitemに紐づいているtransaction_idがすべて削除されたら
+	 *  transactionHistoryからも削除する
+	 * @param transactionId
+	 * 
+	 */
 	@Transactional
-		public void deleteTransactionHistory(Long transactionId) {
-		transactionItemDao.deleteByTransactionId(transactionId);
-		transactionHistoryDao.deleteById(transactionId);	
+		public void deleteTransactionItemAndHistory(Long id) {
+		
+		TransactionItem item = transactionItemDao.findById(id).orElse(null);
+		// IDに該当するデータが存在しない場合は何もしない
+		if (item == null) {
+	        return; 
+	    }
+
+		Long transactionId = item.getTransactionId();
+		
+		// 選択itemを削除
+		transactionItemDao.deleteById(id);
+		
+		
+		boolean hasItems =transactionItemDao.existsByTransactionId(transactionId);
+		
+		 if (!hasItems) {
+		        transactionHistoryDao.deleteById(transactionId);
+		    }
 	}
 }
